pyproject.toml
[tool.poetry]
name = "uidefectmocker"
version = "0.1.0"
description = ""
authors = ["sophiesg <sjiahui27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"
pillow = "^11.0.0"
lxml = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
